requirements.txt
streamlit==1.55.0
langgraph==1.1.3
langchain-openai==1.1.11
langchain-google-genai==4.2.1
langchain-community==0.4.1
langchain-text-splitters==1.1.1
faiss-cpu==1.13.2
ddgs==9.11.3
python-dotenv==1.2.2
tiktoken>=0.9,<1.0
